Constraints, Primary keys, Tables and small fixes
 Constraints now is an abstract class and each constraint type has it's
own class
 Added Support for primary keys with multiple columns, and a class for
them, one for general keys and one for keys that are only one column
 Added implimentetion of different types of tables, and Online table
that requires a connection each time it is accessed, and two Cached
ones, that connect at initialization and copies all rows into memory

 + Added AutoIncrementConstraint, CheckConstraint, DefaultConstraint,
ForeignKeyConstraint, IndexConstraint, NotNullConstraint,
PrimaryKeyConstraint, and UniqueConstraint classes, each extending from
the now abstract, Constraint class, and each contains a getSqlString()
method that returns a string representation of the constriant
 + Added IPrimaryKey, PrimaryKey, SingularPrimaryKey, IBasicPrimaryKey,
BasicPrimaryKey, and SingularBasicPrimaryKey classes that represent
different types of primary keys. they represent respectively: an
interface for primary keys, it's implementation, a version for keys that
consist of only one column, an interface for a basic version of the key
with only values(no columns or cells), it's implementation, and a
version for keys that only have one value
 + Added OnlineTable, CacheTable<T>, and SingularPrimaryKeyCacheTable<T>
table imlplimentations, that represent respectively: a table that
connects to the database each time it's accessed, a table that reads
from memory, and writes to the database, and a version of the previous
one for table that only have one primary key column(improved efficiency)
 + Added RowIterator for converting an Iterator<T extends IRow> into
Iterator<IRow>
 + Added clone(), equals(), and hashCode() to almost every class in
the project

 AbstarctDatabase: Added abstract clone()
 AbstractTable: Added name, primary key constraint, and and indices
fields. Also added them to constructors. Implemented getName(),
getPrimaryKey(), and getIndices() methods, and added abstract clone()
 Column: Added clone(), equals() and hashCode()
 Constraint: Removed info, and target fields, since inheriting classes
that need additional info implement it on their own, and target was
useless on most constraints. Added absract getSqlString() method that
represents the the text version of the constraint, and added equals(), and
hashCode() methods
 ConstraintsEnum: Fixed type, renamed PRIMERY_KEY into PRIMARY_KEY
 DatabaseCell: Fixed VERY stupid mistake of 'column = column' instead of
'this.column = column' in constructor, nobody saw that. Added clone()
hashCode() and equals()
 DatabaseValue: Added clone(), equals(), and hashCode()
 IColumn, IDatabase, IDatabaseHelper: Added clone()
 IDatabaseReader: Added skip() method to skip over next value without
returning it
 IRow: Added getCellsCount(), getPrimaryKey(), hasPrimaryKey(), and
getValue() based on index, on an IColumn object and a column name. Added
clone()
 ITable: Added getPrimaryKey(), getIndices(), clone(), and equals().
Changed getRow to take an IPrimaryKey instead of an IColumn and a
DatabaseValue
 Row: Added primary key field pkey, that gets initialized by searching
the cells for columns with the PRIMARY_KEY constraint. Added
getCellsCount(), getValue(int index), getValue(IColumn), getValue(String
column), getPrimaryKey(), hasPrimaryKey(), clone(), equals(), and
hashCode()